avoid sequential blocking requests in front of the LCP chain
